requirements.txt
Django>=4.1
djangorestframework
psycopg2-binary
channels
channels_redis
drf-yasg 
daphne
